move Experiment
in helper/main
better since it is called/handles a case in main()
